requirements.txt
ipykernel
pydantic
langchain
python-dotenv
langchain-core
langchain_core
langchain-community
langchain_community
pypdf
bs4
arxiv
pymupdf
youtube-transcript-api
pytube
GitPython
docx2txt
wikipedia
langchain-text-splitters
langchain-openai
chromadb
groq
langchain-groq
langchain_nomic 
nomic
sentence_transformers
langchain-huggingface
faiss-cpu
chroma
langchain_chroma
langchain_openai
beautifulsoup4
langchain_ollama
streamlit